# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: npython - Python automation workflow platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows/{id}/packages:
    get:
      operationId: listWorkflowPackages
      tags: [packages]
      summary: List installed packages for a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Package list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: installPackage
      tags: [packages]
      summary: Install a Python package into workflow venv
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallPackageBody"
      responses:
        "200":
          description: Package installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /workflows/{id}/packages/{packageName}:
    delete:
      operationId: uninstallPackage
      tags: [packages]
      summary: Uninstall a Python package from workflow venv
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: packageName
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Package uninstalled
...
  /workflows/{workflowId}/nodes:
    get:
      operationId: listNodes
      tags: [nodes]
      summary: List nodes in a workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
    post:
      operationId: createNode
      tags: [nodes]
      summary: Add a node to a workflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNodeBody"
      responses:
        "201":
          description: Node created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
  /workflows/{workflowId}/nodes/{nodeId}:
    put:
      operationId: updateNode
      tags: [nodes]
      summary: Update a node
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNodeBody"
      responses:
        "200":
          description: Updated node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
    delete:
      operationId: deleteNode
      tags: [nodes]
      summary: Delete a node
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /workflows/{workflowId}/nodes/{nodeId}/execute:
    post:
      operationId: executeNode
      tags: [nodes]
      summary: Execute a single node in isolation
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteNodeBody"
      responses:
        "200":
          description: Node execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeExecutionResult"

  # ── Executions ─────────────────────────────────────────────────────────────
...
  /variables:
    get:
      operationId: listVariables
      tags: [variables]
      summary: List all global variables
      responses:
        "200":
          description: Variable list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Variable"
    post:
      operationId: createVariable
      tags: [variables]
      summary: Create a global variable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVariableBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variable"
  /variables/{id}:
    put:
      operationId: updateVariable
      tags: [variables]
      summary: Update a global variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVariableBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Variable"
    delete:
      operationId: deleteVariable
      tags: [variables]
      summary: Delete a global variable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Credentials ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Node:
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        type:
          type: string
          enum: [code, set_variable, get_variable, trigger_manual, trigger_schedule, condition, http_request, wait, note]
        label:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        config:
          type: object
          description: Node-type specific configuration (code, schedule, condition, etc.)
          additionalProperties: true
        retryCount:
          type: integer
          default: 0
        retryDelayMs:
          type: integer
          default: 1000
        continueOnError:
          type: boolean
          default: false
        stopOnError:
          type: boolean
          default: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, workflowId, type, label, positionX, positionY, config, createdAt, updatedAt]
...
    CreateNodeBody:
      type: object
      properties:
        type:
          type: string
          enum: [code, set_variable, get_variable, trigger_manual, trigger_schedule, condition, http_request, wait, note]
        label:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        config:
          type: object
          additionalProperties: true
      required: [type, label, positionX, positionY]
    UpdateNodeBody:
      type: object
      properties:
        label:
          type: string
        positionX:
          type: number
        positionY:
          type: number
        config:
          type: object
          additionalProperties: true
        retryCount:
          type: integer
        retryDelayMs:
          type: integer
        continueOnError:
          type: boolean
        stopOnError:
          type: boolean
    ExecuteNodeBody:
      type: object
      properties:
        inputData:
          type: object
          additionalProperties: true
    NodeExecutionResult:
      type: object
      properties:
        success:
          type: boolean
        output:
          type: string
        returnValue:
          type: object
          nullable: true
          additionalProperties: true
        durationMs:
          type: integer
        error:
          type: string
          nullable: true
      required: [success, output, durationMs]

    # ── Execution ───────────────────────────────────────────────────────────
...
    Variable:
      type: object
      properties:
        id:
          type: string
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [string, number, boolean, json]
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, key, value, type, createdAt, updatedAt]
    CreateVariableBody:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [string, number, boolean, json]
        description:
          type: string
      required: [key, value, type]
    UpdateVariableBody:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [string, number, boolean, json]
        description:
          type: string

    # ── Credential ──────────────────────────────────────────────────────────
...
    Package:
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        name:
          type: string
        version:
          type: string
          nullable: true
        installedAt:
          type: string
      required: [id, workflowId, name, installedAt]
    InstallPackageBody:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
      required: [name]